추후에 수정용

+++ src/main/webapp/WEB-INF/jsp/usr/article/detail.jsp
@@ -38,16 +38,12 @@
 			<button class="hover:underline" type="button" onclick="history.back();">뒤로가기</button>
 			<c:if test="${article.userCanModify }">
 				<a href="../article/modify?id=${article.id }">수정</a>
-
 			</c:if>
-
-			<c:if test="${article.userCanDelete}">
-				<a href="../article/doDelete?id=${article.id }">삭제</a>
-
+			<c:if test="${article.userCanDelete }">
+				<a onclick="if(confirm('정말 삭제하시겠습니까?') == false) return false;" href="../article/doDelete?id=${article.id }">삭제</a>
 			</c:if>
 
 		</div>
-	</div>
 	</div>
 </section>
